# api/docs/swagger.yaml
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.Alert:
    properties:
      alert_config_id:
        type: string
      body:
        type: string
      create_date:
        type: string
      id:
        type: string
      instruments:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfigInstrument'
        type: array
      name:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      read:
        type: boolean
    type: object
  github_com_USACE_instrumentation-api_api_internal_model.AlertConfig:
    properties:
      alert_email_subscriptions:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.EmailAutocompleteResult'
        type: array
      alert_type:
        type: string
      alert_type_id:
        type: string
      body:
        type: string
      create_date:
        type: string
      creator:
        type: string
      creator_username:
        type: string
      id:
        type: string
      instruments:
        items:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfigInstrument'
        type: array
      last_checked:
        type: string
      last_reminded:
        type: string
      mute_consecutive_alerts:
        type: boolean
      name:
        type: string
      project_id:
        type: string
      project_name:
        type: string
      remind_interval:
        type: string
      schedule_interval:
        type: string
      start_date:
        type: string
      update_date:
        type: string
      updater:
        type: string
      updater_username:
        type: string
      warning_interval:
        type: string
    type: object
  github_com_USACE_instrumentation-api_api_internal_model.AlertConfigInstrument:
    properties:
      instrument_id:
        type: string
      instrument_name:
        type: string
    type: object
...
  github_com_USACE_instrumentation-api_api_internal_model.EmailAutocompleteResult:
    properties:
      email:
        type: string
      id:
        type: string
      user_type:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Monitoring Instrumentation Data Acquisition Systems (MIDAS) Web API
  license:
    name: MIT
    url: https://github.com/USACE/instrumentation-api/blob/555ea51191ff1245fe5910a295862be7514aaec6/LICENSE.md
  title: MIDAS Web API
  version: "2.0"
paths:
...
  /my_alerts:
    get:
      description: list all alerts a profile is subscribed to
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Alert'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists subscribed alerts for a single user
      tags:
      - alert
  /my_alerts/{alert_id}/read:
    post:
      description: |-
        marks an alert as read for a profile
        returning the updated alert
      parameters:
      - description: alert uuid
        format: uuid
        in: path
        name: alert_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Alert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: marks an alert as read
      tags:
      - alert
  /my_alerts/{alert_id}/unread:
    post:
      description: |-
        marks an alert as unread based on provided profile ID and alert ID.
        returning the updated alert
      parameters:
      - description: alert uuid
        format: uuid
        in: path
        name: alert_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Alert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: marks an alert as unread for a profile
      tags:
      - alert
...
  /projects/{project_id}/alert_configs:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists alert configs for a project
      tags:
      - alert-config
    post:
      consumes:
      - application/json
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: alert config payload
        in: body
        name: alert_config
        required: true
        schema:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: creates one alert config
      tags:
      - alert-config
  /projects/{project_id}/alert_configs/{alert_config_id}:
    delete:
      parameters:
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: alert_config_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: deletes an alert config
      tags:
      - alert-config
    get:
      parameters:
      - description: alert config uuid
        format: uuid
        in: path
        name: alert_config_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: gets a single alert
      tags:
      - alert-config
    put:
      consumes:
      - application/json
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: alert config uuid
        format: uuid
        in: path
        name: alert_config_id
        required: true
        type: string
      - description: alert config payload
        in: body
        name: alert_config
        required: true
        schema:
          $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: updates an existing alert config
      tags:
      - alert-config
...
  /projects/{project_id}/instruments/{instrument_id}/alert_configs:
    get:
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.AlertConfig'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists alerts for a single instrument
      tags:
      - alert-config
...
  /projects/{project_id}/instruments/{instrument_id}/alerts:
    get:
      description: list all alerts associated an instrument
      parameters:
      - description: project uuid
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: instrument uuid
        format: uuid
        in: path
        name: instrument_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_USACE_instrumentation-api_api_internal_model.Alert'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: lists alerts for a single instrument
      tags:
      - alert
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"